Always set hide_segment in init_ai_object
Fixes bug where in the editor, you insert a robot with normal behaviour
then play the level - it fails the check on the hide_segment index in
init_ai_objects(). This is because hide_segment was written as the
poisoned value of 0xfdfd if the behaviour of the robot is AIB_NORMAL -
because hide_segment wasn't initialised. Therefore, always setting
hide_segment to suppress this exception (it's an inexpensive operation
seldom called).