render: rewrite how render objects are collected
Increasing the object limit can make this sometimes crash, endless loop,
and corrupt memory... I don't exactly understand how it works, so just
rewrite it. It's somewhat possible that this could happen with the
normal object limit, so maybe there's a general benefit. (Or maybe it
happened only when overflowing the "reserve" list, which would make it
less likely.)