gameseq: autoselect last pilot on program start
Seems convenient, and selecting a pilot was annoying.

I need the additional parameter to RegisterPlayer(), because I want to
both autoselect at program start, and allow selection via menu later.
(Or maybe I could have autoselected if no player was selected yet?)

I noticed there's a command line option to select a player when adding
this, but that is not as convenient and discoverable anyway.