Anti C++ statement
I don't know what kind of funny drugs upstream is doing. The
C++ification makes code unreadable and an extremely ugly mixture of
"modern idiomatic C++" (which in itself is unreadable and ugly), and
ancient C. I don't know how anyone could find this sane or good.

This also applies to the d1x unification, which makes the ifdef hell
even worse than it is, and you have to constantly wonder if some code is
"similar" or not.