sdl: do not grab keys globally if game is paused or not active
There's really no reason to do so. It's nicer if global shortcuts (like
volume control etc.) work if the game is paused.

I abused Screen_mode to detect whether the game is running. This seems
to be the best indicator (without having to change or cleanup FUCKING
EVERYTHING).

Not sure if this attempts to regrab keys in other situations where we
might want that. I spotted the key config dialog, which explicitly
reenables grabbing.