Fix broken orientation matrixes leading to offset object rendering
Some levels contain objects whose initial orientation matrix is slightly
broken. In these cases, the vectors making up the matrix are not fully
orthogonal. In some cases, this leads to a very visible perspective
shift of the rendered object: it appears in the wrong place, and the
shift depends on the viewer position (for example, the object appears to
move sidewards as you increase the distance to it). The problem fixes
itself as soon as a robot notices you (presumably because it modifies
its orientation).

Fix this by calling an existing function that "fixes" the orientation.
This is done when starting a level, not in any other situations
including loading saved games. Games that were saved before this change
will still have the broken matrixes. I'm not fully convinced about what
the function does, but it seems to work well enough.

This is most noticable in the "Entropy Experiment" mission, level 2,
with the blue robot in the upper right "cavern" (when you fly out of the
start room, through the right side tunnel, then up/back). Easier to
observe if you're cloaked and the robots don't "notice" you.