Fix that level is sometimes not renderered from certain viewpoints
Sometimes, when the player was at very specific positions and viewing
angles, the level wasn't rendered. The window size also mattered in
this. Most time, this was only noticeable as rare "flickers", because
you tended to "fly though" this point. Usually it's very hard to
reproduce this bug in a way the screen is permanently black.

This can for example be observed with Descent 2 level 2, at 4K
resolution, after starting the level and flying forward a little bit, in
a position slightly before the middle segment of the start room. On the
map, the ship "sphere" will overlap with the outline of the segment
border.

It turns out this is an overflow problem. The projected screen points
approach INT_MAX, because the viewer is so close to the segment borders,
yet the borders are offscreen. When adding this large value to Canv_w2,
it can overflow, wrapping to a value close to INT_MIN. The renderer
logic sees that the segment "window" boundaries are way offscreen, and
considers the segment invisible. Since the segment window is the only
thing on the screen, the renderer stops with the segment the player is
in, which results in no walls being rendered.

Fix this by checking the addition for overflow. Going through int64_t
and then checking the result is the simplest way (even if lame).