paging: don't unnecessarily preprocess texmerges
Since fvi.c does not require merged textures anymore, and the renderer
pre-merges textures in cases where it's needed anyway, this reduces the
number of texmerges on loading without adding more runtime merges.

In Entropy level 3, this reduces the merges on loading from 290 to 2.
Opening/closing doors still causes runtime merges, but apparently that
was the case already before.