Fix a case of stack corruption
The modified function forgets to remove the current canvas, which was
allocated as stack object. This could cause random overwriting of the
stack after the function returned, but the canvas functions were still
accessing it. This bug must have existed for 10 years or longer.

The reason it became a problem now was that restoring a savegame during
playing mostly did not work, while it worked perfectly fine from the
start screen.