Disallow objects to be located in D2X-XL skybox
The skybox is another bad design decision by D2X-XL. This is a very
large cube that encompasses most of the level, and is used to render an
"outside" view (by using transparent textures where you're supposed to
see the sky).

The problem is that physics sometimes goes crazy and slips objects into
this skybox cube. (Arguably that in itself is an annoying bug, but
finding and fixing this bug is harder.) Admittedly, I have no idea how
this even happens, but it did 1 or 2 times. Try to forbid that objects
move into the skybox, and move them out if they're in the skybox.

(I don't understand why D2X-XL just treated the sky as special case,
maybe make it a cubemap. Maybe it even does, but come on, storing it as
segment? In the normal segment array?)