gamerend: fix another out of bounds read
I don't know why this happens. A debug tool (x-sanitizer) found this.

Disallowing the out of bounds access yields correct results (i.e. no
visual changes). I suspect it has to do with that hires stuff, that was
probably hacked in later, and whoever did that didn't consider the
CM_FULL_SCREEN case, and it happened to work.