robot: fix incorrect robot info circle_distance reads
HOLY FUCKING SHIT WHAT THE FLYING FUCK

The circle_distance[] array should obviously be indexed with j, not i. j
is bounded by the size of the array (NDL define), while i is the index
of the robot_info struct into the Robot_info[] array.

That this is definitely a bug can be seen in the original descent 2
source code release as well. As you can see, this wasn't broken back
then:

https://github.com/OpenSourcedGames/Descent-2/blob/master/SOURCE/MAIN/BM.C#L221

(Although this code is run only for mac - on PC, this read the entire
robot info struct with a single cfread() call.)

D2X-XL doesn't have this bug either. DXX fixed it in the meantime (not
sure when).

This must have been broken before 2006, at which point the imported
history ends (d2x used CVS before that, but didn't import the history).
It's too bad I'll never know which idiot did this and why.

The impact was that 1. circle_distance[] was never correct (affects AI
behavior), and 2. that it overwrote random memory locations with crap.
Basically, HOLY FUCKING SHIT WHAT THE FUCK???