ogl, gamefont: fix window resizing behavior while in menu
Menus are basically a "blocking" function (drive their own event loop
until the menu is left). They keep references to fonts, but resizing the
window recreated the font. This could lead to invalid memory accesses.

"Fix" this by always loading the most high resolution font. I'm not
going to play this game below a resolution of 640x480 ever, so this is
fine. We still need to adjust the font scaling if the window size
changes, but this is ok.