state: don't read past string when saving mission name
This tried to write a fixed size string field with one fwrite call.
Except the string didn't necessarily have that many bytes allocated.
This would be fine if the string buffer were a fixed size array with
enough data, but in this case the string was strdup'ed.

Write it "properly". This stops severely upsetting debug tools.

While we're at it, move this code to separate functions. It's likely
useful for a few other things.