Redefine -autodemo to control enabling of demo playback after inactivity
Before this, -autodemo meant starting the program would start with
playing a demo. This commit essentially removes this feature, but
changes the meaning of the option to control whether demo playback
should start after a timeout. This commit also disables this demo
playback after timeout by default, and requires you to pass this option
to enable it.