Pass correct object to check_effect_blowup
`check_effect_blowup` should receive the `laser_info` of the weapon that
caused the blast.  Previously, it was given the `laser_info` of the
parent of that weapon.  The parent was not of type `OBJ_WEAPON`, so
passing its `laser_info` is meaningless.

Fixes: 9bd1ba7c4748a77d1e8dab0d898f79bedba22eac

(wm4: hopefully equivalent. Also the commit ID Kp cites if the initial
d2x commit (except a different hash - different svn import?), not sure
what's the purpose of that. Removed RELEASE-NOTES.txt entry:
+* Fixed original game bug that could cause a test for a trigger to
produce the wrong result.
)