game: restore grab hack
This restores the behavior of my earlier hack. It seems it broke because
of the "window" abstraction, and Screen_mode is not changed anymore on
pausing.

It's weird that there's no pausing flag, so I'm using the timer flag for
this. It means the grab is not set during automap, but this is probably
OK. Would be trivial to change anyway. To get consistency, update the
grab if the timer is started/stopped.

Definitely this should be folded into the "native" grab handling. It was
written for SDL1, and the API calls were changed in SDL2, so I commented
it. It would be trivial to fix; maybe later.