Remove some unused legacy non-source code files
Not all, just some obvious ones. Some are full of merge conflict
markers, because that was easier when cherry-picking old changes. (What
the hell is up with the convention of duplicating svn and even git
commit messages in CHANGELOG.txt, and to make a new header for them if
it's a new day? This is so disgusting.)