automap: fix unsmooth movement
Movement in the automap was not smooth, but felt slightly "steppy". This
is a regression some time after dxx 0.55.

Something in the automap event refactoring must have broken this. In
fact, it was broken with the commit for free movement. (Not referenced
here because I'm too fucking lazy. Why are you reading this anyway? Go
do something fun instead.) It appears the main cause is that the
movement processing is not run every frame, which of course means the
position is not updated every frame. The idle function is normally run
every frame, except when I hold a key down. SDL2 require you to filter
repeated key events yourself, so maybe the extra key events make the
event code not send an idle event.

It can be fixed by processing the key movement in the draw function.
Since the movement functions essentially interpolate the movement over
time, this should be the correct fix for this.