Remove gross and unnecessary longjmp()
Ewwwwwwwwwwwwwwwwwwwwwww.

This doesn't seem to be necessary - not even with the editor, although I
didn't test that.

It seems dxx already removed this on 2013, I just skipped the
appropriate commit (it probably conflicted and I thought I can do this
myself).

checker.h is something ancient, extremely gross, and fortunately unused.