kconfig: run key ramping exactly once per frame
The key ramping code is not exactly idempotent. If you call it multiple
times in the same frame, it will act differently. This is easily seen in
how adjust_ramped_keyboard_field_div() changes down_time, depending on
the FrameTime. This is a value set per frame, not per kconfig controls
run. (It also changes *time, which references e.g. Controls.pitch_time,
but it's currently reset per kconfig controls invocation anyway.)

Most time, it shouldn't matter, because frames are small and the outcome
of hitting multiple keys in a frame are not that bad, because  there are
only up/down events, no auto repeat. But it's still an issue, and it
annoys me how it seems to depend on the very fine details of when the
idle event is returned.

The mouse_delta_time thing was some hidden state. Move it to Controls
for consistency and because splitting the function requires it to move
anyway.

The automap code does the same dumb crap with temporarily moving its
control_info state (uh, whatever).

Technically, automap will now be invoked 1 frame too late.
ReadControls() really should be changed to be called exactly once per
frame too. This will have to be done later.