kconfig: restore old keyboard behavior
This is about restoring the frame rate dependent key "ramping" behavior
that was present in dxx 0.55.1 and before (possibly the original Descent
versions too on a superficial look at the code, but I didn't test). The
dxx devs put a lot of effort into fixing it, but were off in minor
details. (Though there were other issues which added severe confusion
to this.)

According to my tests, the old code has a 1 frame delay until it reacts.
This is because it tries to take the key down time into account, and the
down time is 0 when the key down state is processed for the first time.
If this is really the case (I didn't test with a version before my
codemonkereyed SDL1->SDL2 port), then this invalidates measurements
about how many frames it takes to get to full speed.

Going by calculations and actual measurements, the old behavior can be
reproduced by setting sensitivity to 18 (out of range), and enforcing a
1 frame delay by using the previous r->down_time value each frame. This
is for pitch/heading, other axes are instantaneous after the 1 frame
delay.

Another weird issue is that the new code does not reset movement
correctly when hitting opposing keys. You can get a small movement even
if both opposing keys are held down. I suspect this is because the new
code doesn't clamp r->down_time correctly. As soon as down_time is >=
F1_0, it can remain being > F1_0. If the opposing keys end up at
different down_time values, they won't compensate each others. It
doesn't happen with the old code, because it sets r->down_time to a
constant value. Of course this would be easy to fix.

A third issue is that the key ramping was not correctly reset in some
situations. If you hit e.g. the left key and the slide-on key, and
you release the slide-on key, the ramping for the left key (i.e.
changing the heading now) should start again. Not doing this created
the impression that ship navigation was "slippery" when slide-on was
used. This is pretty important to me. The new code has this problem
because the down_time field is shared between these two cases, even
though the two cases require different ramping. The "new old" code
doesn't, because down_time is either 0 or F1_0 (the ramping is
instead essentially done on the target axis, with the correct scale).

For now, I decided to backport the old code, and enable it under a user
option (which is enabled by default, because I'm probably the only
user). The old key_down_time[] array isn't needed, because using
FrameTime essentially gives the same results. I'm also keeping the very
weird special handling of heading/pitch, which essentially leads to
faster stopping (I think...). For other axes, there's essentially no
ramping.

The dxx semantics, which seem to match current upstream, are unchanged.
I think this is better and less risky than trying to mess with the new
code, and also should allow easier testing of the differences. It's also
not too intrusive (the new control_type enum and associated parameter,
which is the most intrusive part of the commit, could be used to
simplify the code to a better state than before this commit).

Testing this is pretty simple. Add this to control.c:

	if (test_start) {
		static int frame;
		frame++;
		printf("%d: ", frame);
		printf("fw=%d ", forward_thrust_time);
		printf("pi=%d ", Controls.pitch_time);
		printf("hd=%d ", Controls.heading_time);
		printf("bk=%d ", Controls.bank_time);
		printf("sd=%d ", Controls.sideways_thrust_time);
		printf("ve=%d ", Controls.vertical_thrust_time);
		printf("\n");
	}

and set test_start to true as soon as you get a key down event from SDL.
The changes to do so are too different across the 3 versions tested (my
0.55 fork, my updated fork, dxx upstream). Hit a key and keep it pressed
for a second, then plot the printed values (gnuplot works, changing the
output to make it easier consumable by gnuplot helps).