state: fix undefined behavior
This just dumped random parts of the stack into the savegame. ASAN
complained. Change the (apparently correct) read code from a seek to
byte reads too.

(DXX broke and then fixed it. The fix was on my merge backlog, but I'm
doing it differently because I'm not comfortable with seeks.)