Separate input event processing from per frame processing
This is like the earlier kconfig split. This essentially relied on
getting IDLE events, which in my SDL2 port apparently don't come
necessarily every frame. You could argue that it was probably intended
to be sent every frame, but it's a bad idea anyway.

Fixes weird energy->shields conversion speed.