window: move undefined behavior from a dozen source files to a function
window_exists() always returns 1, unless you rely on undefined behavior
(or you don't use window_create() to allocate the window). C does not
allow using the value of a pointer to deallocated memory for anything.
For a good reason: a window allocated in the meantime could use exactly
the same pointer value.

Unfortunately, window_exists() has a lot of callers. Fortunately, most
of them just want to run the event loop until a window goes away, so
this code can be moved to a new window_run_event_loop() function, where
the problem can hopefully be fixed (even if not, it's a nice
simplification). Unfortunately, the menu code and the event code itself
contain pretty tricky uses of it. Fucking shit.

The window system was added within the last 10 years, i.e. it's pretty
new code for an old code base like this. I thought it was a pretty clean
and nice change, but actually it's full of issues caused by utter
incompetence and not thinking things though. What do you think have GUI
toolkits tried in the last 30 years? They tried to get less crashy by
avoiding moronic dumb shit like this.