Redo key grab handling (yet again)
Same semantics, but slightly cleaner. This makes use of the window
system to keep track whether to globally grab keys or not. Considering
how pause works now (after the dxx cherry-picks), this is probably
better. It's also good not to rely on the odd Screen_mode.

Make the automap grab for now. It makes sense because grabbing is
supposed to allow using system keys for normal movement commands.