Simplify switching between menus, game, etc.
Until now, the game window was explicitly hidden with
window_set_visible() calls. Likewise, menus were explicitly hidden or
reshown.

Simplify this by letting the window stack handle this. There is no
reason to e.g. hide the game window, if another window covers it anyway.
There is actually no reason to explicitly hide windows at all, so remove
the window visibility flag. Replace it by a flag that tells the window
code whether a window is opaque. (Originally I wanted to use the canvas
size, but it's not reliable for this purpose. E.g. the main menu does
not cover the whole screen, even though it draw over the whole screen.)

We don't need modal windows either. Remove them.

The whole show_menus()/hide_menus() circus becomes simply unnecessary.

One problem is that the automap could prevent that the game is paused.
Apparently this is not correct in multiplayer (makes sense).
game_handler has checks for this anyway, so maybe it's not a problem.
But on the other hand, EVENT_WINDOW_DRAW and thus GameProcessFrame will
not be called anymore, which could be a problem. I have no way of
testing this. If it is a problem, there should be a EVENT_FRAME or such,
that handles such per-frame tasks that need to be performed even if the
window is invisible.

newmenu.c contains a gross hack. The issue is that some menus are not
opaque (like the pause window), while others are (like the end game
scores, secret level transfer messages). It's hard to pass new
parameters because there's a weird fuckup of multiple newmenu_* entry
points with tons of arguments that call each others, so I preferred the
gross hack. It can be cleaned up later.

The loop in inferno.c is unnecessary; the loop above it exits only if
window_get_front() returns NULL. Besides, this loop didn't consider
invisible windows.

I'm obviously not caring much about the editor code.