gamecntl: prevent freespace cheat from banking on next level
The "e" key is normally set to "bank right". If you enter "freespace",
it will show a dialog, and then jump to the level you enter.

This fixes that the game thought "e" was pressed on the next level. It
happened because the "e" down event was propagated to the kconfig code
_after_ the dialog asking for the new level returned. The "e" up event
was probably incorrectly dispatched to the dialog, but since the down
event is fed to it at a later time (due to reentrant event loops), it
didn't matter anyway.

Reporting the triggering "e" down event as consumed fixes this.

An alternative fix would be avoiding reentrant event loops, but this
would probably require a lot of refactoring.

This needs to be done for other cheat codes that enter reentrant event
loops. Probably best to do for all cheats, but not now.