weapon: rewrite weapon cycling code
This code was rewritten with a recent cherry-pick, but for some reason
it cycled through non-autoselect weapons, which is definitely not what I
want. (Or maybe it's a bug, either in the original commit, or due to
cherry-picking.)

Rewrite it again, and ignore non-autoselectable weapons. Remove the
shitty code duplication too. This code should be flexible enough to
easily change behavior to anything else.

This includes trying to handle the fucked up super laser case correctly.
It seems the original dxx 0.55.1 code simply ignored the super laser
entry in the autoselect order.