Remove u_mem wrappers
Considering modern debugging tools, this stuff is more in the way than
it helps. Also, the possibility of confusing pointers allocated with the
wrappers and standard functions exists.

Remove the wrapper. Make the existing functions aliases (i.e. use the
old !NDEBUG case) to avoid having to change all the source code.