ogl: restore 3D text rendering
Conversion of the coordinates from fix to float was missing. (Though the
old code was missing that too, so why did it work?)

Also we can use GL_TRIANGLE_FAN, which with 4 vertices apparently
generates the same triangles (assuming the vertices are in the same
plane) as GL_QUADS. So we can avoid GL_QUADS, which is even more legacy
than all the other GL API usage in this code.