Replace some bitshifts with multiplication/division
The multiplication is turned into a straight shift by most compilers.
The division generates slightly longer code, but it's not too bad.

Most importantly, the change from shift to division changes the rounding
behavior. But this should be OK too.

The goal is avoiding undefined behavior with shifts, such as shifts of
negative numbers.