render: change rendering order of transparent/alpha sides
The Descent 2 visibility/depth sorting algorithm is quite imperfect. It
does a breadth-first search from the viewer's segment, renders the
segments in reverse order, and calls it a day. This doesn't always work:
e.g. in Descent 2 level 4, move close to the first force field after the
spawn position, observe how part of the left tunnel is visible in front
of the walls that should cover it. Fortunately, the Z-buffer saves you
here.

This doesn't work with alpha blended surfaces. You need to render these
in the right order. I made some attempts to fix the depth sorting. One
idea was using the fact that if you find a wall that leads to a segment
that 1. should be after the current segment, and 2. was already visited,
it means the depth search made a mistake, which you can fix by removing
the segment from the result list and searching again. This didn't work
out, because triangulation means 2 segments don't necessarily have a
strict order (i.e. unfixable unless you sort triangles), and also it
revisited too often.

However, I still had to notice that dxx 0.55.1 handled a certain case
much better than current dxx master (including the changes I merged). In
mission Area 51, there is a block with permanently half-cloaked walls
(i.e. rendered as flat poly with alpha), just in the room on the right
of the start position. If you look onto it from the top, there's a
"slit" that lets you see the "inside" without an alpha texture over it.
In this position, the "inside" texture tends to be considered before the
also visible alpha sides of the block that border on the "slit", even
though it should be behind it.

The newer dxx upstream code makes this pretty ugly, while dxx 0.55.1 is
not perfect, but reasonable. This commit restores the old behavior for
better or worse. Maybe worse, as the "fix" for this is just a hack, that
most likely will have to be reverted anyway because it might make more
important level parts look worse.

The trick is to _not_ write to the Z-buffer when rendering alpha
textures, like the newer dxx code explicitly did. If you do, you will
hide objects and other walls even though they should be visible through
them. Rendering the object before the alpha texture when it should be
behind it is better than making the object invisible for the parts where
it should be visible. Same for alpha walls interacting with each other:
better blend them in the wrong order than hiding one of them.

The Z-buffer is still read, because if an opaque texture covers the
alpha texture, the alpha texture is indeed discarded.

Also change the code that attempted to treat "almost opaque" parts of an
alpha texture. Make it actually render only opaque pixels in this
situation. Likewise, the second pass renders only translucent pixels.
(At least this is the intention. I didn't test it much.)

Results would be slightly better if we'd do the same with object
textures, but that's really where this hack starts failing big time.

DXX upstream introduced the behavior change with "Draw laser effects
with special blending [...]" (d020a3136e5 in master). The removal of the
glDepthMask(GL_FALSE); call is essentially reverted here, but some
details are different.