vecmat: remove float usage in fix codepath
The fuck????????

fixmuldiv is (fix)((((fix64)a)*b)/c). It shouldn't be able to overflow,
because fix is a 32 bit type, fix64 is 64 bits, and the result of a 32
bit multiplication fits in 64 bit. It's possible that the final result
(after the division) doesn't fit into fix, but the float code has the
same problem.

I don't know why "DPH" (whoever that is?) thought this would overflow.
Either I'm overlooking something, or maybe fixmuldiv() used to be buggy.
Relative to the age of the codebase, fast 64 bit integers are rather
new, and the fixmuldiv() before that was probably terrible.