render: handle d2x-xl sky segment differently and remove duplicated code
Originally I did this for simplicity. But then a commit merged from dxx
changed the rendering code (and in fact removed render_segment()). I
kept the old function for the sky segment (I guess). Fix this.

On a side note, the sky segment is an excellent example how the Descent
search algorithm can fail to sort segments by depth. The sky box is
normally the segment with the greatest depth at any point within normal
level geometry. Yet the algorithm would just append it to the result
list, implying that any further segments are _behind_ it. The Z-buffer
will save you, but it's still nicer to strictly append it to the list
after depth search is done. (Also, you can't add the sky segment as
child, because sort_seg_children() requires direct children.)