render: rotate points eagerly
The code tries to go through a lot of effort to reduce vertex rotations.
but you can observe that if a segment gets rendered, its vertices are
rotated anyway. Likewise, having at least 1 child means all segment
vertices are rotated and projected. So change the code to always rotate
visited segments, and always project vertices if there's a child.

Also, if the segment is off the screen, don't render it by setting its
Render_list entry to -1.

Reduces code complexity and the number of rotate_list() calls in the
tested levels. The actual number of points rotated seems to be about the
same.

We could go further and use a rotation matrix in GL, and rotate segments
only if the visibility search algorithm needs it. But currently, it
always needs it anyway, so there's not much of a point.