render: remove pointless visited[] array
It seems this was only needed for window_check==0. In release mode, this
was always 1, and all code associated with window_check==0 was removed
recently. visited[] is a leftover which now does nothing.

The only other thing visited[] could do is preventing rendering segments
more than once. But this is already guaranteed as Render_list does not
contain duplicate entries (other than -1, which means render nothing).