render: avoid another potential out of bounds access
lcnt can already be MAX_RENDER_SEGS (i.e. out of bounds) when accessing
render_windows[]. Obviously, the intention was to directly write at this
position for the sake of the segment addition code path, while it only
served as temporary storage for the segment revisit path. So store it on
the stack instead, and copy the window to the destination as needed.

One weird thing is the "Render_list[lcnt] = -1;" line. lcnt is the first
free index (or points to the end of the array), so this must be a dead
write. This confuses me in addition, because all code seemed to assume
that Render_list can contain -1 entries, while this is the last -1 store
from the original code. Remove this apparently redundant and possibly
out of bounds store. (Dealing with -1 Render_list entries is still
needed, because I added another -1 store when a segment is outside of
the screen.)