render: reduce number of redundant iterations
The search loop reiterated over the entire result list in each depth
iteration. This happened because scnt was reset to 0 on start of the
search loop. (The "scnt = ecnt;" at the end of the loop is a dead
store.)

Use instead the lowest index of any segment that needs to be
reprocessed.

This brings down the total number of per-segment iterations from ten
thousands of iterations to hundreds of iterations. In one tested case it
went from ~50k down to ~350 iterations. Oops. (Although most iterations
immediately skipped to the next element after checking the processed[]
array, which is linearly read in addition. So the actual performance
improvement will not be as big as it sounds.)