render: more cosmetic improvements to visibility search loop
Mostly stuff like merging variable declaration and initialization, or
renaming variables.

lcnt is removed, and N_render_segs is used directly instead.

Remove the goto. It was fine, but gotos are ugly. The loop terminates
slightly later due to this change (it cannot add new segments, so it
stops if there are no new segments to process).