render: restore not visiting children whose sides are offscreen
This reverts an accidental removal of the same logic, when rotate_list
was moved up to always be run on a segment. I basically moved the check
up to perform it per block, but that doesn't really make much sense.
It's better to keep it per child, which will not visit offscreen blocks
anyway.