wall, render: simplify wall_is_doorway and add alpha information
Attempt to remove some of the code duplication from the function. The
main observation that can be made here is that if WID_RENDER_FLAG is
given, the texture transparency flag normally need to be checked.

For some of these, there is the theoretical chance that weird level data
relies on the specifics how WALL_DOOR_OPENED interacts with the wall
type, so this behavior is awkwardly kept. I don't know if there are
levels that use this, or if there's anything in the rest of the code
that prevents this situation in the first place.

There are some more special cases where assumptions about transparency
change. We now always rely on the texture flags. If game data exists
that uses incorrectly opaque flagged textures for WALL_BLASTABLE, or
WALL_DOOR with state==WALL_DOOR_OPENING, then it might be rendered
incorrectly.

Unfortunately, it turned out that I can't really use the alpha
information for anything. If bilinear filtering is on, even binary
transparency produces arbitrary alpha values. Objects are rendered after
_all_ opaque geometry has been rendered, so objects won't be visible
under the filtered alpha gradients, which looks like shit. So treating
binary and alpha transparency differently doesn't help with anything,
and the new flag is pointless. Well, whatever.