rand: remove "watcom" (?) code and fix declaration
The code under NO_WATCOM_RAND uses the C stdlib, but it's never built.
Maybe the Descent developers did use Watcom C, so the code that is used
today was written or copied by d2x to make sure it has the same
behavior.

Anyway, I guess having a deterministic/portable implementation of it is
justified and maybe required, so keep the currently active code.

The source file didn't include maths.h, which has the prototype for the
defined functions. Fix this. Also fix the fact that d_rand was declared
incorrectly. Because the C standard committee is retarded and dumb, ()
means "unspecified parameters", and you need to use (void). (Not sure
what is more disturbing, that the C committee didn't fix it yet, or
working with a C code base full of such obvious mistakes. I can only
hope that some ancient C compilers/dialects actually had differing
semantics.)