Fix sound menu and build with SDL_mixer
During my messy cherry-picks, I decided to merge the jukebox changes.
But this was a mistake. It's some incredibly messy code that somehow
tries to make the game an audio player, and (most annoyingly) used new
messy PhysFS APIs to iterate directories (including having a file
browser UI in the sound config menu). The latter parts made me comment
the entire sound config menu during cherry-picking, with the intent to
fix or remove it later. The sdl_mixer code is a scons option and was
disabled by default, so I never happened to test it, but it was broken.

Remove all the jukebox stuff. You can instead use a music player in the
background (this is what I do). Fix the sdl_mixer code. Remove redbook
CD audio (it used SDL support, but SDL2 removed support). Reenable the
sound menu, and remove the unsupported options.

Movie sound seems broken in sdl_mixer mode. I couldn't test
non-sdl_mixer mode much because SDL is a damn POS and refused to accept
the descent settings on my system.