Remove "Screen Resolution" menu
SDL2 fortunately discouraged the awful practice of changing the screen
resolution when going to fullscreen. They support a proper "desktop"
fullscreen mode, which is the only thing this code base supports after
it was converted to SDL2. Consequently, screen resolution setting was
without function.

Remove the screen resolution configuration, because I don't intend to
support resolution switching. You can do this with a 3rd party tool
anyway. Note that the config file entry is still used for remembering
the windowed size (but it doesn't need GUI to set this).

The aspect configuration was also ignored, which could be considered a
bug. But on the other hand all it did for me in the past was
accidentally playing the game in a non-square aspect ratio, so remove
this as well.

The fullscreen checkbox is still vaguely useful (especially since there
are DEs that don't give you obvious ways to change it), so move it to
the "Graphics Options" menu.

Unfortunately, this still overwrites the window size when going into
fullscreen; this needs to be fixed later. Also, the menu entry for the
removed sub menu is replaced by a placeholder, because of the awful menu
code, which also can be fixed later.