guages: fix an instance of idiotic string handling
s was initialized with memset (and duplicating the array length of
course), when you you have initialized it with "".

strncpy() is completely pointless here; could have used memcpy(), which
wouldn't have made the compiler bark (because the strncpy() looks unsafe
here; it relies on the memset() before).

snprintf() is called with the size argument set to the source string
size instead of the buffer size. It was even explicitly changed from
sprintf(). This is of course no safer than sprintf, and extremely
misleading. If the callsign string is longer than CALLSIGN_LEN+10, the
buffer would overflow. (In this case, callsign is guaranteed to be
shorter, so you could claim the code looks worse than it is.)

And the best part of this: you don't need a separate buffer at all.

(How can you get it so wrong?)

Untested, so I might fuck it up too.