gauges: remove broken gr_bitmap() calls in copckpit code
Remove 2 gr_bitmap() calls and all code that leads to them. Assert()
instead if this happens.

The gr_bitmap() calls are broken, because they imply a screen to screen
transfer. The calls use a canvas bitmap as source, that is a
sub-canvas/sub-bitmap from the screen. The screen is a GL framebuffer,
so the bitmap is just unused dummy data. There's nothing in gr_bitmap()
and friends that properly deals with the situation. This code base, dxx
55.1, and dxx master all don't do anything useful when explicitly
triggering this code path. (You only get crashes or attempts to use
extremely slow per-pixel copies. Granted, these tests might be with out
of bounds coordinates.)

To do this correctly, you'd probably have to use FBOs or so.

I didn't manage to trigger these code paths. All HUD/cockpit modes seem
to be fine. I have no idea what this code is trying to do either. So
just remove them, and Assert() on these situations instead. If this ever
happens, the fix would likely look different to this code anyway.